+++ pkg/common/common.go
@@ -14,6 +14,8 @@ import (
 const (
 	// ReadUser is the global user for reading logs
 	ReadUser = "read"
+	// DefaultWriteTenant is the default tenant for writing logs
+	DefaultWriteTenant = "giantswarm"
 	// Grafana Multi Tenant Proxy Ingress
 	proxyIngressNamespace = "monitoring"
 	proxyIngressName      = "grafana-multi-tenant-proxy"