+++ pkg/resource/proxy-auth/proxy-auth.go
@@ -82,7 +82,7 @@ func GenerateProxyAuthSecret(lc loggedcluster.Interface, credentialsSecret *v1.S
 			Password: writePassword,
 			// we set the tenant even though it may be given by the sender (promtail)
 			// depending of grafana-multi-teant-proxy config
-			Orgid: writeUser,
+			Orgid: "giantswarm",
 		})
 
 		// Add write user to allowed tenants for read user